pyproject.toml
[tool.poetry]
name = "conversation-chatbot"
version = "0.1.0"
description = ""
authors = ["Karakchiev Aleksandr <realrushen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.11"
environs = "^9.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
